# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Trivias y Preguntas",
        "description": "Documentación de la API para la aplicación de Trivias y Preguntas.",
        "contact": {
            "email": "test@email.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor de Desarrollo"
        }
    ],
    "paths": {
...
        "/trivias/crear": {
            "post": {
                "tags": [
                    "Trivias"
                ],
                "summary": "Crea una nueva trivia",
                "operationId": "cf3b09dac1e896092267510f4756c4a4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "numero_trivia",
                                    "nombre_trivia",
                                    "pregunta_ids"
                                ],
                                "properties": {
                                    "numero_trivia": {
                                        "description": "Número identificador de la trivia",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nombre_trivia": {
                                        "description": "Nombre descriptivo de la trivia",
                                        "type": "string",
                                        "example": "Trivia de Recursos Humanos"
                                    },
                                    "pregunta_ids": {
                                        "type": "array",
                                        "items": {
                                            "description": "Lista de IDs de preguntas asociadas a la trivia (MongoDB ObjectId)",
                                            "type": "string",
                                            "example": "68387d483208ef498505f092"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Trivia creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Trivia creada con éxito."
                                        },
                                        "trivia": {
                                            "description": "Detalles de la trivia creada",
                                            "properties": {
                                                "_id": {
                                                    "type": "string",
                                                    "example": "68387d483208ef498505f093"
                                                },
                                                "numero_trivia": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nombre_trivia": {
                                                    "type": "string",
                                                    "example": "Trivia de Recursos Humanos"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflicto: Trivia ya existe",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ya existe una trivia con este número de trivia."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validación fallida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Los datos proporcionados no son válidos."
                                        },
                                        "errors": {
                                            "description": "Detalles de los errores de validación",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error interno del servidor."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/trivias/listar": {
            "get": {
                "tags": [
                    "Trivias"
                ],
                "summary": "Listar trivias o una trivia por número",
                "operationId": "bbba0888b97f6421b5432db6dc85029e",
                "parameters": [
                    {
                        "name": "numero_trivia",
                        "in": "query",
                        "description": "Número de la trivia a consultar (opcional). Si se omite, lista todas las trivias.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de trivias o trivia específica encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "_id": {
                                                "description": "ID único de la trivia (MongoDB ObjectId)",
                                                "type": "string",
                                                "example": "68387d483208ef498505f093"
                                            },
                                            "numero_trivia": {
                                                "description": "Número de la trivia",
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "nombre_trivia": {
                                                "description": "Nombre de la trivia",
                                                "type": "string",
                                                "example": "Trivia de Recursos Humanos"
                                            },
                                            "pregunta_ids": {
                                                "description": "IDs de preguntas asociadas",
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                },
                                                "example": [
                                                    "68387d483208ef498505f092"
                                                ]
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "example": [
                                    {
                                        "_id": "6838e481574f16179305e352",
                                        "numero_trivia": 1,
                                        "nombre_trivia": "Trivia de Historia",
                                        "pregunta_ids": [
                                            "6838f481574f16179305e353"
                                        ]
                                    },
                                    {
                                        "_id": "6838e481574f16179305e354",
                                        "numero_trivia": 2,
                                        "nombre_trivia": "Trivia de Ciencias",
                                        "pregunta_ids": []
                                    }
                                ]
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "No Content: No se encontró la trivia específica (si se consultó por número y no existe)."
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error interno del servidor."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingresa el token Bearer (ej. 'Bearer your_token_here')",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}